Hangman-Python
The program picks a random word from the list. The user tries to know this. The number of letters of the word to be guessed is given to the user. He has four rights to know the word. If there is an input other than a letter, the program warns the user. If the said letter is in the word, every correctly guessed letter in that word is shown to the user.